% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer-exit.R
\name{exit_frame}
\alias{exit_frame}
\title{Find frame for exit handlers}
\usage{
exit_frame(envir, frames = as.list(sys.frames()), calls = as.list(sys.calls()))
}
\description{
\itemize{
\item If knitr is running and \code{defer()} is run at top-level in a chunk,
\code{exit_frame()} returns the first frame on the stack that inherits
from the knitr namespace. This way handlers are run when knitr
returns.
\item If \code{getOption("withr.hook_source")} is \code{TRUE}, and \code{defer()} is
run at top-level in a script being evaluated by \code{\link[base:source]{base::source()}},
\code{exit_frame()} returns the frame of \code{source()}. This way handlers
are run when the script has finished running.
}
}
\keyword{internal}
