% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer.R
\name{global_defer}
\alias{global_defer}
\title{Defer expression globally}
\usage{
global_defer(expr, priority = c("first", "last"))
}
\arguments{
\item{expr}{\verb{[expression]}\cr An expression to be evaluated.}

\item{priority}{\verb{[character(1)]}\cr Specify whether this handler should
be executed \code{"first"} or \code{"last"}, relative to any other
registered handlers on this environment.}
}
\description{
This function is mostly internal. It is exported to be called in
standalone \code{defer()} implementations to defer expressions from the
global environment.
}
\keyword{internal}
