% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_.R
\name{scope_}
\alias{scope_}
\title{Create a new "scope" function}
\usage{
scope_(set, reset = set, envir = parent.frame())
}
\arguments{
\item{set}{\code{[function(...)]}\cr Function used to set the state.
The function can have arbirarily many arguments, they will be replicated
in the formals of the returned function.}

\item{reset}{\code{[function(x)]}\cr Function used to reset the state.
The first argument can be named arbitrarily, further arguments with default
values, or a "dots" argument, are supported but not used: The function will
be called as \code{reset(old)}.}

\item{envir}{\code{[environment]}\cr Environment of the returned function.}
}
\value{
\code{[function(new, code, ...)]} A function with at least two arguments,
\itemize{
\item \code{new}: New state to use
\item \code{code}: Code to run in that state.
}
If there are more arguments to the function passed in \code{set} they are
added to the returned function.  If \code{set} does not have arguments,
the returned function only has a \code{code} argument.
}
\description{
This function is a "constructor" for \code{scope_...} functions.  It
is only needed if you want to alter some global state which is
not covered by the existing \code{scope_...} functions, see
\link{withr-package} for an overview.
}
\keyword{internal}

