% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/devices.R
\name{with_bmp}
\alias{with_bmp}
\alias{with_cairo_pdf}
\alias{with_cairo_ps}
\alias{with_dev}
\alias{with_device}
\alias{with_jpeg}
\alias{with_pdf}
\alias{with_postscript}
\alias{with_svg}
\alias{with_tiff}
\alias{with_xfig}
\title{Graphics devices}
\usage{
with_bmp(new, code, width = 480, height = 480, units = "px",
  pointsize = 12, bg = "white", res = NA, family = "sans",
  restoreConsole = TRUE, type = c("windows", "cairo"),
  antialias = c("default", "none", "cleartype", "gray", "subpixel"))

with_cairo_pdf(new, code, width = 7, height = 7, pointsize = 12,
  onefile = FALSE, family = "sans", bg = "white",
  antialias = c("default", "none", "gray", "subpixel"),
  fallback_resolution = 300)

with_cairo_ps(new, code, width = 7, height = 7, pointsize = 12,
  onefile = FALSE, family = "sans", bg = "white",
  antialias = c("default", "none", "gray", "subpixel"),
  fallback_resolution = 300)

with_jpeg(new, code, width = 480, height = 480, units = "px",
  pointsize = 12, quality = 75, bg = "white", res = NA,
  family = "sans", restoreConsole = TRUE, type = c("windows", "cairo"),
  antialias = c("default", "none", "cleartype", "gray", "subpixel"))

with_pdf(new, code, width, height, onefile, family, title, fonts, version,
  paper, encoding, bg, fg, pointsize, pagecentre, colormodel, useDingbats,
  useKerning, fillOddEven, compress)

with_postscript(new, code, onefile, family, title, fonts, encoding, bg, fg,
  width, height, horizontal, pointsize, paper, pagecentre, print.it, command,
  colormodel, useKerning, fillOddEven)

with_svg(new, code, width = 7, height = 7, pointsize = 12,
  onefile = FALSE, family = "sans", bg = "white",
  antialias = c("default", "none", "gray", "subpixel"))

with_tiff(new, code, width = 480, height = 480, units = "px",
  pointsize = 12, compression = c("none", "rle", "lzw", "jpeg", "zip",
  "lzw+p", "zip+p"), bg = "white", res = NA, family = "sans",
  restoreConsole = TRUE, type = c("windows", "cairo"),
  antialias = c("default", "none", "cleartype", "gray", "subpixel"))

with_xfig(new, code, onefile = FALSE, encoding = "none",
  paper = "default", horizontal = TRUE, width = 0, height = 0,
  family = "Helvetica", pointsize = 12, bg = "transparent",
  fg = "black", pagecentre = TRUE, defaultfont = FALSE,
  textspecial = FALSE)
}
\arguments{
\item{new}{\code{[named character]}\cr New graphics device}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temporarily use a graphics device.
}
\examples{
with_pdf(c(file = "test.pdf", width = 7, height = 5), plot(runif(5)))
}
\seealso{
\code{\link{withr}} for examples

\code{\link[grDevices]{Devices}}
}

