% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/env.R
\name{with_envvar}
\alias{with_envvar}
\title{Environmental Variables}
\usage{
with_envvar(new, code, action = "replace")
}
\arguments{
\item{new}{\code{[any]}\cr Values for setting}

\item{code}{\code{[any]}\cr Code to execute in that environment}

\item{action}{should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing variables with the same name.}
}
\description{
Environmental Variables
}
\details{
if \code{NA} is used those #' environment variables will be unset.
If there are any duplicated variable #' names only the last one is used.
}

