% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torture.R
\name{with_gctorture2}
\alias{with_gctorture2}
\title{Torture Garbage Collector}
\usage{
with_gctorture2(new, code, wait = new, inhibit_release = FALSE)
}
\arguments{
\item{new}{\verb{[integer]}\cr run GC every 'step' allocations.}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{wait}{integer; number of allocations to wait before starting
    \abbr{GC} torture.}

\item{inhibit_release}{logical; do not release free objects for
    re-use: use with caution.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
argument.
}
\description{
Temporarily turn gctorture2 on.
}
\seealso{
\code{\link{withr}} for examples
}
