% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lib.R
\name{with_libpaths}
\alias{with_libpaths}
\title{library paths, replacing current libpaths}
\usage{
with_libpaths(new, code, action = "replace")
}
\arguments{
\item{new}{Values for setting}

\item{code}{Code to execute in that environment}

\item{action}{should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing paths.}
}
\description{
library paths, replacing current libpaths
}

