% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{with_namespace}
\alias{with_namespace}
\title{Execute code with a modifed search path}
\usage{
with_namespace(namespace, code)
}
\arguments{
\item{namespace}{\code{[character(1)]}\cr Namespace / package name to load.}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
This attaches a namespace to the search path, executes the code, then
removes the package from the search path. The package namespace is _not_
unloaded however.
}
\examples{
\dontrun{
with_namespace("ggplot2", {
  ggplot(mtcars) + geom_point(aes(wt, hp))
})
}
}
\seealso{
\code{\link{withr}} for examples
}
