% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/par.R
\name{with_par}
\alias{with_par}
\title{Graphics Parameters}
\usage{
with_par(new, code, no.readonly = FALSE)
}
\arguments{
\item{new}{Values for setting}

\item{code}{Code to execute in that environment}

\item{no.readonly}{see \code{\link{par}} documentation.}
}
\description{
Graphics Parameters
}

