% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{with_seed}
\alias{with_seed}
\alias{with_preserve_seed}
\title{Random seed}
\usage{
with_seed(seed, code)

with_preserve_seed(code)
}
\arguments{
\item{seed}{`[integer(1)]`\cr The random seed to use to evaluate the code.
Pass `NA` to use the RNG to draw the random seed, pass `NULL` to not change
the random seed at all (and also to avoid resetting it).}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
`with_seed()` runs code with a specific random seed and resets it afterwards.

`with_preserve_seed()` runs code with the current random seed and resets it
  afterwards.
}
\seealso{
\code{\link{withr}} for examples
}
