% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{with_trace}
\alias{with_trace}
\alias{local_trace}
\title{Debugging states which unset themselves}
\usage{
with_trace(
  what,
  code,
  tracer = NULL,
  exit = NULL,
  at = numeric(),
  print = TRUE,
  signature = NULL,
  where = topenv(parent.frame()),
  edit = FALSE
)

local_trace(
  what,
  code,
  tracer = NULL,
  exit = NULL,
  at = numeric(),
  print = TRUE,
  signature = NULL,
  where = topenv(.local_envir),
  edit = FALSE,
  .local_envir = parent.frame()
)
}
\arguments{
\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{.local_envir}{\verb{[environment]}\cr The environment to use for scoping.}

\item{file, .file}{\verb{[named list]}\cr Files to create.}

\item{...}{Additional (possibly named) arguments of files to create.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
argument.
}
\description{
Create debugging states, which are then automatically unset afterwards.
}
\examples{
with_file("file1", {
  writeLines("foo", "file1")
  readLines("file1")
})

with_file(list("file1" = writeLines("foo", "file1")), {
  readLines("file1")
})

}
\seealso{
\code{\link{withr}} for examples
}
