% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\alias{set_path}
\title{Get/set the PATH variable.}
\usage{
set_path(path, action = "prefix")
}
\arguments{
\item{path}{character vector of paths}

\item{action}{One of \code{"prefix"} (default), \code{"suffix"} or \code{"replace"},
determines if the existing paths should be prepended, appended, or replaced by the
existing paths.}
}
\value{
\code{set_path} invisibly returns the old path.
}
\description{
Get/set the PATH variable.
}
\seealso{
\code{\link{with_path}} to temporarily set the path for a block
  of code

Other path: \code{\link{with_path}}
}

