% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/with_.R
\name{with_}
\alias{with_}
\title{Create a new "with" function}
\usage{
with_(set, reset = set, envir = parent.frame())
}
\arguments{
\item{set}{Function used to set the state.}

\item{reset}{Function used to reset the state.}

\item{envir}{Environment of the returned function.}
}
\value{
\code{with_()} A function with at least two arguments,
\itemize{
\item \code{new}: New state to use
\item \code{code}: Code to run in that state.
}
If there are more arguments to the function passed in \code{set} they are
added to the returned function.  If \code{set} does not have arguments,
the returned function only has a \code{code} argument.
}
\description{
This function is a "constructor" for \code{with_...} functions.  This
function is only needed if you want to alter some global state which is
not covered by the existing \code{with_...} functions, see
\link{withr-package} for an overview.
}
\keyword{internal}

