% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collate.r
\name{with_collate}
\alias{with_collate}
\title{Collation Order}
\usage{
with_collate(new, code)
}
\arguments{
\item{new}{\code{[any]}\cr Values for setting}

\item{code}{\code{[any]}\cr Code to execute in that environment}
}
\description{
Collation Order
}

