% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collate.r
\name{with_collate}
\alias{with_collate}
\title{Collation Order}
\usage{
with_collate(new, code)
}
\arguments{
\item{new}{\code{[character(1)]}\cr New collation order}

\item{code}{\code{[any]}\cr Code to execute in that environment}
}
\description{
Temporarily change collation order by changing the value of the
\code{LC_COLLATE} locale.
}

