% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{with_connection}
\alias{with_connection}
\title{Automatically close connections}
\usage{
with_connection(...)
}
\arguments{
\item{...}{One or more named parameters with the connection(s) to create and
one unnamed parameter with the code to run.}

\item{con}{The connection to create, this is returned by
\code{local_connection()}.}
}
\description{
Connections which are automatically closed.
}
\examples{
with_connection(con = file("foo", "w"), {
  writeLines(c("foo", "bar"), con)
})

read_foo <- function() {
  readLines(local_connection(file("foo", "r")))
}
read_foo()
}
