% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/working_directory.R
\name{with_dir}
\alias{in_dir}
\alias{with_dir}
\title{Working Directory}
\usage{
with_dir(new, code)

in_dir(new, code)
}
\arguments{
\item{new}{Values for setting}

\item{code}{Code to execute in that environment}
}
\description{
Working Directory
}

