% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dir.R
\name{with_dir}
\alias{with_dir}
\title{Working directory}
\usage{
with_dir(new, code)
}
\arguments{
\item{new}{\code{[character(1)]}\cr New working directory}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temorarily change the current working directory.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{setwd}}
}

