% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpaths.R
\name{with_libpaths}
\alias{local_libpaths}
\alias{with_libpaths}
\title{Library paths}
\usage{
with_libpaths(new, code, action = "replace")

local_libpaths(new, code, action = "replace",
  .local_envir = parent.frame())
}
\arguments{
\item{new}{\code{[character]}\cr New library paths}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{action}{\code{[character(1)]}\cr should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing paths.}

\item{.local_envir}{\code{[environment]}\cr The environment to use for scoping.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temporarily change library paths.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{.libPaths}}

Other libpaths: \code{\link{with_temp_libpaths}}
}

