% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{with_locale}
\alias{scope_locale}
\alias{with_locale}
\title{Locale settings}
\usage{
with_locale(new, code)

scope_locale(new, code, .scoped_envir = parent.frame())
}
\arguments{
\item{new}{\code{[named character]}\cr New locale settings}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{.scoped_envir}{\code{[environment]}\cr The environment to use for scoping.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temporarily change locale settings.
}
\details{
Setting the \code{LC_ALL} category is currently not implemented.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{Sys.setlocale}}
}

