% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locale.R
\name{with_locale}
\alias{with_locale}
\title{any locale setting}
\usage{
with_locale(new, code)
}
\arguments{
\item{new}{\code{[any]}\cr Values for setting}

\item{code}{\code{[any]}\cr Code to execute in that environment}
}
\description{
any locale setting
}

