% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/par.R
\name{with_par}
\alias{with_par}
\title{Graphics parameters}
\usage{
with_par(new, code, no.readonly = FALSE)
}
\arguments{
\item{new}{\code{[named list]}\cr New graphics parameters and their values}

\item{code}{\code{[any]}\cr Code to execute in that environment}

\item{no.readonly}{\code{[logical(1)]}\cr see \code{\link{par}} documentation.}
}
\description{
Temporarily change graphics parameters.
}
\seealso{
\code{\link{par}}
}

