% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.R
\name{with_path}
\alias{with_path}
\title{PATH environment variable}
\usage{
with_path(new, code, action = "prefix")
}
\arguments{
\item{new}{Values for setting}

\item{code}{Code to execute in that environment}

\item{action}{should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing paths.}
}
\description{
PATH environment variable
}
\seealso{
Other path: \code{\link{path}}, \code{\link{set_path}}
}

