% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.R
\name{with_path}
\alias{with_path}
\title{PATH environment variable}
\usage{
with_path(new, code, action = "prefix")
}
\arguments{
\item{new}{\code{[character]}\cr New \code{PATH} entries}

\item{code}{\code{[any]}\cr Code to execute in that environment}

\item{action}{\code{[character(1)]}\cr Should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing paths}
}
\description{
Temporarily change the system search path.
}
\seealso{
\code{\link{Sys.setenv}}
}

