% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{with_path}
\alias{scoped_path}
\alias{with_path}
\title{PATH environment variable}
\usage{
with_path(new, code, action = "prefix")

scoped_path(new, code, action = "prefix", .scoped_envir = parent.frame())
}
\arguments{
\item{new}{\code{[character]}\cr New \code{PATH} entries}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{action}{\code{[character(1)]}\cr Should new values \code{"replace"}, \code{"prefix"} or
\code{"suffix"} existing paths}

\item{.scoped_envir}{\code{[environment]}\cr The environment to use for scoping.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temporarily change the system search path.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{Sys.setenv}}
}

