% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{with_sink}
\alias{with_sink}
\title{Output redirection}
\usage{
with_sink(new, code, append = FALSE, type = c("output", "message"),
  split = FALSE)
}
\arguments{
\item{new}{a writable \link{connection} or a character string naming the file
to write to, or \code{NULL} to stop sink-ing.}

\item{code}{\code{[any]}\cr Code to execute in the temporary environment}

\item{append}{logical.  If \code{TRUE}, output will be appended to
    \code{file}; otherwise, it will overwrite the contents of
    \code{file}.}

\item{type}{character string.  Either the output stream or the messages
    stream.  The name will be partially matched so can be abbreviated.}

\item{split}{logical: if \code{TRUE}, output will be sent to the new
    sink and to the current output stream, like the Unix program \code{tee}.}
}
\value{
\code{[any]}\cr The results of the evaluation of the \code{code}
  argument.
}
\description{
Temporarily divert output to a file via \code{\link{sink}}.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{sink}}
}

