% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libpaths.R
\name{with_temp_libpaths}
\alias{with_temp_libpaths}
\title{Library paths}
\usage{
with_temp_libpaths(code)
}
\arguments{
\item{code}{\code{[any]}\cr Code to execute in that environment}
}
\description{
Temporarily prepend a new temporary directory to the library paths.
}
\seealso{
\code{\link{withr}} for examples

\code{\link{.libPaths}}

Other libpaths: \code{\link{with_libpaths}}
}

